// src/swapi/models/PlanetRecord.ts
export interface PlanetRecord {
  url:             string;
  name:            string;
  rotation_period: string;
  orbital_period:  string;
  diameter:        string;
  climate:         string;
  gravity:         string;
  terrain:         string;
  surface_water:   string;
  population:      string;
}